refactor: move core files to internal/ilc package - Renamed and relocated core files to `internal/ilc` for better encapsulation - Updated `main.go` to use the new `ilc` package structure - Adjusted imports and references to reflect the new package organization
