+++ .gitignore
@@ -15,3 +15,4 @@
 # vendor/
 
 dist/
+ilc
