+++ internal/ilc/history.go
@@ -1,4 +1,4 @@
-package main
+package ilc
 
 import (
 	"errors"
