+++ internal/ilc/runner_test.go
@@ -1,4 +1,4 @@
-package main
+package ilc
 
 import (
 	"testing"
