+++ internal/ilc/yaml.go
@@ -1,4 +1,4 @@
-package main
+package ilc
 
 import (
 	"fmt"
